vote/requirements.txt
Flask
Redis
gunicorn
google-cloud-pubsub